requirements.txt
langchain==0.3.0
langchain-core>=0.1.0
langchain-community>=0.0.15
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
aiohttp>=3.9.0
firecrawl-py>=0.1.0 
tavily-python>=0.1.0